% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_calc.R
\name{pivot_calc}
\alias{pivot_calc}
\title{Title}
\usage{
pivot_calc(data, rows = NULL, cols = NULL, value = NULL, fun = sum, pivot = T)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.}

\item{rows}{a character vector of items}

\item{cols}{a character vector of items}

\item{pivot}{logical: should wide table be returned - col categories as columns (TRUE), or left long and tidy (FALSE)?}
}
\description{
Title
}
\examples{
flat_titanic \%>\% pivot_calc(rows = sex, value = freq)
flat_titanic \%>\% pivot_calc(rows = sex, fun = mean, value = freq)
}
