% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_count.R
\name{pivot_count}
\alias{pivot_count}
\title{library(tidyverse)
library(magrittr)
create_tidy_titanic() \%>\%
pivot_count(Sex, Survived)}
\usage{
pivot_count(data, cols = NULL, rows = NULL, pivot = T)
}
\description{
library(tidyverse)
library(magrittr)
create_tidy_titanic() \%>\%
pivot_count(Sex, Survived)
}
