% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_helper.R
\name{pivot_helper}
\alias{pivot_helper}
\title{A function that pretty much let's you do anything in the tidypivot space, is carefully crafted, and is adapted to make the other functions easy to use}
\usage{
pivot_helper(
  data,
  rows = NULL,
  cols = NULL,
  value = NULL,
  wt = NULL,
  within = NULL,
  fun = NULL,
  pivot = T,
  wrap = F
)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.}

\item{rows}{a character vector of items}

\item{cols}{a character vector of items}

\item{pivot}{logical: should wide table be returned - col categories as columns (TRUE), or left long and tidy (FALSE)?}
}
\value{

}
\description{
A function that pretty much let's you do anything in the tidypivot space, is carefully crafted, and is adapted to make the other functions easy to use
}
\examples{
tidy_titanic \%>\% pivot_helper(rows = sex, cols = survived, fun = length) # pivot_count
flat_titanic \%>\% pivot_helper(rows = sex, value = freq, fun = mean) # pivot_calc
flat_titanic \%>\% pivot_helper(rows = sex, value = freq, fun = sum) # pivot_count (weighted sum)
nar <- function(x) return(NA)
flat_titanic \%>\% pivot_helper(rows = sex, cols = survived, fun = nar); #pivot_null
sample1 <- function(x) sample(x, 1)
flat_titanic \%>\% pivot_helper(rows = sex, cols = survived, fun = sample1, value = freq); #pivot_sample1
samplen <- function(x, n) paste(sample(x, 5, replace = F), collapse = ", ")
flat_titanic \%>\% pivot_helper(rows = sex, cols = survived, fun = samplen, value = freq); #pivot_samplen
paste_collapse <- function(x) paste (x, collapse = ", ")
flat_titanic \%>\% pivot_helper(rows = sex, fun = paste_collapse, value = freq) #pivot_list
}
