% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_prop.R
\name{pivot_prop}
\alias{pivot_prop}
\title{Title}
\usage{
pivot_prop(
  data,
  rows = NULL,
  cols = NULL,
  value = NULL,
  fun = sum,
  within = NULL,
  pivot = T,
  percent = T,
  round = F
)
}
\arguments{
\item{round}{}
}
\value{

}
\description{
Title
}
