% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piping_demo.R
\name{set_rows}
\alias{set_rows}
\title{Title}
\usage{
set_rows(first, vars)
}
\arguments{
\item{vars}{}
}
\description{
Title
}
\examples{
  tp_init(data = "tidytitanic::tidy_titanic")

set_data(data = tidytitanic::tidy_titanic) \%>\%
 set_rows(age)
}
