% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piping_demo.R
\name{tp_add}
\alias{tp_add}
\title{Title}
\usage{
tp_add(
  nothing = NULL,
  data = NULL,
  rows = NULL,
  cols = NULL,
  fun = NULL,
  value = NULL,
  wt = NULL,
  pivot_logical = NULL,
  quietly = FALSE
)
}
\arguments{
\item{pivot_logical}{}
}
\description{
Title
}
\examples{
tp_init(data = "tidytitanic::tidy_titanic")
  tp_add(rows = "sex")
  tp_add(cols = "class")

tp_init(data = "tidytitanic::tidy_titanic") |>
  tp_add(rows = "sex") |>
  tp_add(cols = "class")

}
