% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piping_demo.R
\name{tp_init}
\alias{tp_init}
\title{Title}
\usage{
tp_init(data = NULL, quietly = FALSE)
}
\arguments{
\item{pivot_logical}{}
}
\description{
Title
}
\examples{
tp_init(data = "tidytitanic::tidy_titanic")
}
