% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piping_demo.R
\name{user_function_initiate}
\alias{user_function_initiate}
\title{Title}
\usage{
user_function_initiate(
  data = NULL,
  rows = NULL,
  cols = NULL,
  fun = NULL,
  value = NULL,
  wt = NULL,
  pivot_logical = NULL
)
}
\arguments{
\item{pivot_logical}{}
}
\description{
Title
}
\examples{
user_function_initiate()
}
