% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_count.R
\name{pivot_count}
\alias{pivot_count}
\title{Title}
\usage{
pivot_count(data, cols = NULL, rows = NULL, pivot = TRUE, wt = NULL)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.}

\item{cols}{a character vector of items}

\item{rows}{a character vector of items}

\item{pivot}{logical: should wide table be returned - col categories as columns (TRUE), or left long and tidy (FALSE)?}
}
\description{
Title
}
\examples{
tidy_titanic \%>\% pivot_count(rows = sex)
tidy_titanic \%>\% pivot_count(cols = sex)
tidy_titanic \%>\% pivot_count(rows = survived, cols = sex)
tidy_titanic \%>\% pivot_count(rows = c(survived, sex), cols = age)
tidy_titanic \%>\% pivot_count(rows = c(survived), cols = c(age, sex))
tidy_titanic \%>\% pivot_count(cols = c(survived), rows = c(age, sex, class))
tidy_titanic \%>\% pivot_count(rows = c(survived, sex), cols = age, pivot = FALSE)
flat_titanic \%>\% pivot_count(rows = sex, wt = freq)
}
