% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_prop.R
\name{pivot_prop}
\alias{pivot_prop}
\title{Title}
\usage{
pivot_prop(
  data,
  y,
  y0,
  y00,
  x = NULL,
  value = NULL,
  fun = sum,
  within,
  within2,
  pivot = !is.null(x),
  percent = T,
  round = F
)
}
\arguments{
\item{round}{}
}
\value{

}
\description{
Title
}
