% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_prop.R
\name{pivot_prop}
\alias{pivot_prop}
\title{Title}
\usage{
pivot_prop(
  data,
  rows = NULL,
  cols = NULL,
  value = NULL,
  fun = sum,
  within = NULL,
  pivot = T,
  percent = T,
  round = F
)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.}

\item{rows}{a character vector of items}

\item{cols}{a character vector of items}

\item{pivot}{logical: should wide table be returned - col categories as columns (TRUE), or left long and tidy (FALSE)?}
}
\value{

}
\description{
Title
}
\examples{
tidy_titanic \%>\% pivot_prop(rows = sex, cols = survived, within = sex)
}
