% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piping_demo.R
\name{set_cols}
\alias{set_cols}
\title{Title}
\usage{
set_cols(vars)
}
\arguments{
\item{vars}{}
}
\description{
Title
}
\examples{
tidytitanic::tidy_titanic |>
set_data()
set_rows(age)
set_cols(survived)
}
